Fixed code smells and moved integration test file
